+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/text/TextComponentBuilder.kt
@@ -1,5 +1,6 @@
 package io.github.lmliam.kotventure.core.text
 
+import io.github.lmliam.kotventure.core.style.StyleScope
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.TextComponent
 import net.kyori.adventure.text.format.Style
@@ -19,6 +20,10 @@ internal class TextComponentBuilder : TextScope {
 
     override fun style(style: Style) {
         builder.style(style)
+    }
+
+    override fun style(init: StyleScope.() -> Unit) {
+        TextStyleScope(builder).init()
     }
 
     override fun decorate(decoration: TextDecoration) {
@@ -57,6 +62,14 @@ internal class TextComponentBuilder : TextScope {
 
     override fun text(init: TextScope.() -> Unit) {
         builder.append(TextComponentBuilder().apply(init).build())
+    }
+
+    override fun append(component: Component) {
+        builder.append(component)
+    }
+
+    override fun newline() {
+        builder.append(Component.newline())
     }
 
     internal fun build(): Component = builder.build()